only handle deserialization of state, no discriminator or length
